fix(auth): stop asking deployments to remove AUTH_TYPE
Startup warned on every boot whenever AUTH_TYPE was set to disabled or to one
of the retired single-provider modes, telling operators to remove it. Most
deployments cannot: the one-shot seed migration and the legacy SAML import
still read the variable, so a deployment that has not seeded yet needs it.
The result was a permanent warning instructing people to do something unsafe.

Drop the env read. The mode already comes from MULTI_TENANT alone, so the
function now only records which mode is in use and is named for that.

The two migration readers are untouched: they are one-way imports that stop
mattering once an sso_provider row exists, and neither asks an operator to
change config.
